backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
transformers==4.35.2
torch==2.1.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
nltk==3.8.1
openai==1.3.7
requests==2.31.0
python-jose[cryptography]==3.3.0
